extract render_author & render_example from base.py into separate files remove hardcoded meta images from code and always use WorkflowMetadata update grid layout to not render separator on last row use gif preview for displaying output video fix migrations
