+++ daras_ai_v2/meta_preview_url.py
@@ -4,16 +4,26 @@ import typing
 
 from furl import furl
 
+PreviewSizes = typing.Literal[
+    "400x400", "1170x1560", "40x40", "72x72", "80x80", "96x96"
+]
+
+DEFAULT_META_IMG = (
+    # Small
+    "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/ec2100aa-1f6e-11ef-ba0b-02420a000159/thumbs/Main_400x400.jpg"
+    # "https://storage.googleapis.com/dara-c1b52.appspot.com/meta_tag_default_img.jpg"
+    # Big
+    # "https://storage.googleapis.com/dara-c1b52.appspot.com/meta_tag_gif.gif"
+)
+
 
 def meta_preview_url(
     file_url: str | None,
-    fallback_img: str = None,
-    size: typing.Literal[
-        "400x400", "1170x1560", "40x40", "72x72", "80x80", "96x96"
-    ] = "400x400",
-) -> str | None:
+    fallback_img: str = DEFAULT_META_IMG,
+    size: PreviewSizes = "400x400",
+) -> tuple[str | None, bool]:
     if not file_url:
-        return fallback_img
+        return fallback_img, False
 
     f = furl(file_url)
     dir_segments = f.path.segments[:-1]
@@ -23,17 +33,9 @@ def meta_preview_url(
 
     if content_type.startswith("video/"):
         f.path.segments = dir_segments + ["thumbs", f"{base}.gif"]
-        # fallback to default image if video gif not present
-        file_url = fallback_img
-    elif content_type in ["image/png", "image/jpeg", "image/tiff", "image/webp"]:
-        # sizes:  400x400,1170x1560,40x40,72x72,80x80,96x96
+        return str(f), True
+    elif content_type in {"image/png", "image/jpeg", "image/tiff", "image/webp"}:
         f.path.segments = dir_segments + ["thumbs", f"{base}_{size}{ext}"]
-
-    new_url = str(f)
-    return new_url
-    ## this is too costly to do for every api call
-    # r = requests.head(new_url)
-    # if r.status_code == 200:
-    #     return new_url
-    # else:
-    #     return file_url
+        return str(f), False
+    else:
+        return file_url, False
