+++ daras_ai_v2/profiles.py
@@ -1,11 +1,11 @@
 import typing
 from html import escape as escape_html
 
-from django.utils.translation import ngettext
 import gooey_gui as gui
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError, transaction
 from django.db.models import Count
+from django.utils.translation import ngettext
 from fastapi import Request
 from furl import furl
 
@@ -29,6 +29,7 @@ from daras_ai_v2.meta_content import (
 )
 from daras_ai_v2.urls import paginate_queryset, paginate_button
 from handles.models import Handle
+from widgets.saved_workflow import render_saved_workflow_preview
 from workspaces.models import Workspace, WorkspaceMembership
 from workspaces.widgets import set_current_workspace
 
@@ -261,17 +262,13 @@ def render_public_runs_list(request: Request, workspace: Workspace):
 
     def _render(pr: PublishedRun):
         workflow = Workflow(pr.workflow)
-        page_cls = workflow.page_cls
-        page_cls().render_example(
+        render_saved_workflow_preview(
+            workflow.page_cls,
             pr,
             workflow_pill=f"{workflow.get_or_create_metadata().emoji} {workflow.short_title}",
         )
 
-    for pr in prs:
-        _render(pr)
-        # render divder only if not the last item
-        if pr != prs[-1]:
-            gui.div(className="mb-2 mt-2 border-bottom")
+    grid_layout(1, prs, _render)
 
     paginate_button(url=request.url, cursor=cursor)
