+++ daras_ai_v2/utils.py
@@ -1,7 +1,6 @@
-from django.utils import timezone
 from datetime import timedelta, datetime
 
-from bots.models import Workflow
+from django.utils import timezone
 
 THRESHOLDS = [
     (timedelta(days=365), "y"),
