+++ recipes/BulkRunner.py
@@ -35,8 +35,6 @@ from daras_ai_v2.workflow_url_input import (
 from gooeysite.bg_db_conn import get_celery_result_db_safe
 from recipes.DocSearch import render_documents
 
-DEFAULT_BULK_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/d80fd4d8-93fa-11ee-bc13-02420a0001cc/Bulk%20Runner.jpg.png"
-
 
 class BulkRunnerPage(BasePage):
     title = "Bulk Runner"
@@ -91,9 +89,6 @@ _(optional)_ Add one or more Gooey.AI Evaluator Workflows to evaluate the result
 List of URLs to the evaluation runs that you requested.
             """,
         )
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_BULK_META_IMG
 
     def render_form_v2(self):
         gui.write(f"##### {field_title_desc(self.RequestModel, 'run_urls')}")
