+++ recipes/DocSearch.py
@@ -42,8 +42,6 @@ from daras_ai_v2.vector_search import (
 from furl import furl
 from pydantic import BaseModel
 
-DEFAULT_DOC_SEARCH_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/bcc7aa58-93fe-11ee-a083-02420a0001c8/Search%20your%20docs.jpg.png"
-
 
 class DocSearchPage(BasePage):
     PROFIT_CREDITS = 3
@@ -141,9 +139,6 @@ class DocSearchPage(BasePage):
         cache_knowledge_widget(self)
         gui.write("---")
         doc_search_advanced_settings()
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_DOC_SEARCH_META_IMG
 
     def preview_description(self, state: dict) -> str:
         return "Add your PDF, Word, HTML or Text docs, train our AI on them with OpenAI embeddings & vector search and then process results with a GPT3 script. This workflow is perfect for anything NOT in ChatGPT: 250-page compliance PDFs, training manuals, your diary, etc."
