+++ recipes/EmailFaceInpainting.py
@@ -19,8 +19,6 @@ from recipes.FaceInpainting import FaceInpaintingPage
 email_regex = r"\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b"
 twitter_handle_regex = r"(@)?[A-Za-z0-9_]{1,15}"
 
-DEFAULT_EMAIL_FACE_INPAINTING_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/6937427a-9522-11ee-b6d3-02420a0001ea/Email%20photo.jpg.png"
-
 
 class EmailFaceInpaintingPage(FaceInpaintingPage):
     title = "AI Generated Photo from Email Profile Lookup"
@@ -84,9 +82,6 @@ class EmailFaceInpaintingPage(FaceInpaintingPage):
     @classmethod
     def get_example_preferred_fields(self, state: dict) -> list[str]:
         return ["email_address"]
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_EMAIL_FACE_INPAINTING_META_IMG
 
     def preview_description(self, state: dict) -> str:
         return "Find an email's public photo and then draw the face into an AI generated scene using your own prompt + the latest Stable Diffusion or DallE image generator."
