+++ recipes/QRCodeGenerator.py
@@ -42,7 +42,6 @@ from url_shortener.models import ShortenedURL
 from daras_ai_v2.enum_selector_widget import enum_multiselect
 
 ATTEMPTS = 1
-DEFAULT_QR_CODE_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/a679a410-9456-11ee-bd77-02420a0001ce/QR%20Code.jpg.png"
 
 
 class QrSources(Enum):
@@ -123,11 +122,6 @@ class QRCodeGeneratorPage(BasePage):
         raw_images: list[FieldHttpUrl]
         shortened_url: FieldHttpUrl | None
         cleaned_qr_code: FieldHttpUrl
-
-    def preview_image(self, state: dict) -> str | None:
-        if len(state.get("output_images") or []) > 0:
-            return state["output_images"][0]
-        return DEFAULT_QR_CODE_META_IMG
 
     def related_workflows(self) -> list:
         from recipes.CompareText2Img import CompareText2ImgPage
@@ -437,11 +431,6 @@ Here is the final output:
     def render_output(self):
         state = gui.session_state
         self._render_outputs(state)
-
-    def render_example_preview_media(self, published_run: PublishedRun):
-        output_images = list(published_run.saved_run.state.get("output_images", []))
-        if output_images:
-            gui.image(output_images[0])
 
     def render_run_preview_output(self, state: dict):
         self._render_outputs(state, max_count=1)
