+++ recipes/RelatedQnA.py
@@ -16,8 +16,6 @@ from recipes.DocSearch import render_doc_search_step, EmptySearchResults
 from recipes.GoogleGPT import GoogleGPTPage
 from recipes.RelatedQnADoc import render_qna_outputs
 
-DEFAULT_SEO_CONTENT_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/cbd2c94e-9456-11ee-a95e-02420a0001cc/People%20also%20ask.jpg.png"
-
 
 class RelatedGoogleGPTResponse(GoogleGPTPage.ResponseModel):
     search_query: str
@@ -46,9 +44,6 @@ class RelatedQnAPage(BasePage):
     class ResponseModel(BaseModel):
         output_queries: list[RelatedGoogleGPTResponse]
         serp_results: dict
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_SEO_CONTENT_META_IMG
 
     def render_description(self) -> str:
         return "This workflow gets the related queries for your Google search, searches your custom domain and builds answers using the results and GPT."
