+++ recipes/RelatedQnADoc.py
@@ -15,8 +15,6 @@ from daras_ai_v2.vector_search import render_sources_widget
 from recipes.DocSearch import DocSearchPage, render_doc_search_step, EmptySearchResults
 from recipes.GoogleGPT import render_output_with_refs, GoogleSearchMixin
 
-DEFAULT_QNA_DOC_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/bab3dd2a-538c-11ee-920f-02420a00018e/RQnA-doc%20search%201.png.png"
-
 
 class RelatedDocSearchResponse(DocSearchPage.ResponseModel):
     search_query: str
@@ -43,9 +41,6 @@ class RelatedQnADocPage(BasePage):
     class ResponseModel(BaseModel):
         output_queries: list[RelatedDocSearchResponse]
         serp_results: dict
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_QNA_DOC_META_IMG
 
     def render_description(self) -> str:
         return "This workflow gets the related queries for your Google search, searches your custom domain and builds answers using the results and GPT."
