+++ recipes/SocialLookupEmail.py
@@ -24,7 +24,6 @@ from daras_ai_v2.loom_video_widget import youtube_video
 from daras_ai_v2.redis_cache import redis_cache_decorator
 
 email_regex = r"\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b"
-DEFAULT_SOCIAL_LOOKUP_EMAIL_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/6729ea44-9457-11ee-bd77-02420a0001ce/Profile%20look%20up%20gpt%20email.jpg.png"
 
 
 class SocialLookupEmailPage(BasePage):
@@ -82,9 +81,6 @@ class SocialLookupEmailPage(BasePage):
             VideoBotsPage,
             EmailFaceInpaintingPage,
         ]
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_SOCIAL_LOOKUP_EMAIL_META_IMG
 
     def render_description(self):
         gui.write(
