+++ recipes/Text2Audio.py
@@ -15,8 +15,6 @@ from daras_ai_v2.img_model_settings_widgets import (
     num_outputs_setting,
 )
 
-DEFAULT_TEXT2AUDIO_META_IMG = "https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/85cf8ea4-9457-11ee-bd77-02420a0001ce/Text%20guided%20audio.jpg.png"
-
 
 class Text2AudioModels(Enum):
     audio_ldm = "AudioLDM (CVSSP)"
@@ -59,9 +57,6 @@ class Text2AudioPage(BasePage):
             typing.Literal[tuple(e.name for e in Text2AudioModels)],
             list[FieldHttpUrl],
         ]
-
-    def preview_image(self, state: dict) -> str | None:
-        return DEFAULT_TEXT2AUDIO_META_IMG
 
     def render_form_v2(self):
         gui.text_area(
