+++ widgets/author.py
@@ -1,0 +1,110 @@
+import html
+
+import gooey_gui as gui
+
+from app_users.models import AppUser
+from workspaces.models import Workspace
+
+
+def render_author_from_workspace(
+    workspace: Workspace | None,
+    *,
+    image_size: str = "30px",
+    responsive: bool = True,
+    show_as_link: bool = True,
+    text_size: str | None = None,
+) -> object:
+    if not workspace:
+        return
+    photo = workspace.get_photo()
+    if workspace.is_personal:
+        name = workspace.created_by.display_name
+    else:
+        name = workspace.display_name()
+    if show_as_link and workspace.is_personal and workspace.handle_id:
+        link = workspace.handle.get_app_url()
+    else:
+        link = None
+    return render_author(
+        photo=photo,
+        name=name,
+        link=link,
+        image_size=image_size,
+        responsive=responsive,
+        text_size=text_size,
+    )
+
+
+def render_author_from_user(
+    user: AppUser | None,
+    *,
+    image_size: str = "30px",
+    responsive: bool = True,
+    show_as_link: bool = True,
+    text_size: str | None = None,
+):
+    if not user:
+        return
+    photo = user.get_photo()
+    name = user.full_name()
+    if show_as_link and (handle := user.get_handle()):
+        link = handle.get_app_url()
+    else:
+        link = None
+    return render_author(
+        photo=photo,
+        name=name,
+        link=link,
+        image_size=image_size,
+        responsive=responsive,
+        text_size=text_size,
+    )
+
+
+def render_author(
+    photo: str | None,
+    name: str | None,
+    link: str | None,
+    *,
+    image_size: str,
+    responsive: bool,
+    text_size: str | None,
+):
+    if not photo and not name:
+        return
+
+    if responsive:
+        responsive_image_size = f"calc({image_size} * 0.67)"
+    else:
+        responsive_image_size = image_size
+
+    linkto = link and gui.link(to=link) or gui.dummy()
+    with linkto, gui.div(className="d-flex align-items-center"):
+        if photo:
+            with gui.styled(
+                """
+                    @media (min-width: 1024px) {
+                        & {
+                            width: %(image_size)s;
+                            height: %(image_size)s;
+                        }
+                    }
+                    """
+                % dict(image_size=image_size)
+            ):
+                gui.image(
+                    photo,
+                    style=dict(
+                        width=responsive_image_size,
+                        height=responsive_image_size,
+                        marginRight="6px",
+                        borderRadius="50%",
+                        objectFit="cover",
+                        pointerEvents="none",
+                    ),
+                )
+
+        if name:
+            name_style = {"fontSize": text_size} if text_size else {}
+            with gui.tag("span", style=name_style):
+                gui.html(html.escape(name))
